use_prompt_regression/server/requirements.txt
openenv-core
openai
pydantic
fastapi
uvicorn
python-dotenv